+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.2.1]
+### Added
+- Added Support for SMS and TTS only VerifyRequests
+
 ## [5.2.0]
 ### Added
 - Added support for verification workflows inside of `VerifyRequest`.